// src/types/store.ts
export interface FoodOption {
  id: string;
  nombre: string;
  precio_usd: number;
  activo?: boolean;
}

export interface FoodOptionGroup {
  id: string;
  nombre: string;
  min_select: number;
  max_select: number;
  options: FoodOption[];
}